extract common to HBSS and VBSS
